+++ src/routes/docs/classes/+page.svelte
@@ -40,7 +40,7 @@
 		'grid',
 		'float_none|left|right|inline_start|inline_end',
 		'flip_x|y|xy',
-		'font_sans|serif|mono',
+		'font_family_sans|serif|mono',
 		'line_height_xs-xl',
 		'size_xs-xl9',
 		'icon_size_xs-xl3',
@@ -149,7 +149,7 @@ ${'<' as string}script>
 
 <style>
 	.classes {
-		font-family: var(--font_mono);
+		font-family: var(--font_family_mono);
 	}
 	.classes li {
 		padding: var(--space_xs3);