+++ src/routes/docs/colors/Color_Swatch_Item.svelte
@@ -39,7 +39,7 @@
 	li {
 		display: flex;
 		align-items: stretch;
-		font-family: var(--font_mono);
+		font-family: var(--font_family_mono);
 		min-height: var(--input_height_sm);
 	}
 	li:hover {