+++ src/routes/docs/colors/Hue_Swatch.svelte
@@ -30,7 +30,7 @@
 	li {
 		display: flex;
 		align-items: stretch;
-		font-family: var(--font_mono);
+		font-family: var(--font_family_mono);
 		padding: var(--space_xs2);
 	}
 	li:hover {
@@ -73,6 +73,6 @@
 		display: flex;
 	}
 	.description {
-		font-family: var(--font_sans);
+		font-family: var(--font_family_sans);
 	}
 </style>